refactor: move storage code into storage/ subdirectory
Consolidate storage.rs, disk_storage.rs, and storage-related items from
lib.rs into a storage/ module directory with shared trait, types, and
per-backend files. Make BucketState private in both backends.
